requirements.txt
alembic==1.4.2
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coverage==5.1
cycler==0.10.0
Cython==0.29.20
dnspython==1.16.0
email-validator==1.1.1
fastapi==0.54.1
Flask==1.1.2
google-cloud-pubsub==1.6.1
gunicorn==20.0.4
h11==0.9.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
kiwisolver==1.2.0
Mako==1.1.2
MarkupSafe==1.1.1
matplotlib==3.2.1
more-itertools==8.3.0
numpy==1.18.4
packaging==20.4
pandas==1.0.3
pluggy==0.13.1
psycopg2==2.8.5
py==1.8.1
pydantic==1.5.1
pydantic[email]
pymongo==3.10.1
pyparsing==2.4.7
pytest==5.4.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
requests==2.23.0
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
six==1.14.0
SQLAlchemy==1.3.16
SQLAlchemy-Paginator==0.1
starlette==0.13.2
threadpoolctl==2.1.0
urllib3==1.25.9
uvicorn==0.11.5
wcwidth==0.1.9
websockets==8.1
Werkzeug==1.0.1
